+++ tfhe/src/integer/server_key/radix_parallel/mod.rs
@@ -20,6 +20,8 @@ mod shift;
 mod sub;
 
 #[cfg(test)]
+pub(crate) mod tests_cases_comparisons;
+#[cfg(test)]
 pub(crate) mod tests_cases_unsigned;
 #[cfg(test)]
 mod tests_signed;
